requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --all-extras --output-file=requirements.txt pyproject.toml
#
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via
    #   frictionless
    #   jsonlines
    #   jsonschema
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
cfgv==3.4.0
    # via pre-commit
chardet==5.2.0
    # via frictionless
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer
coverage[toml]==7.4.4
    # via pytest-cov
cryptography==42.0.5
    # via pyjwt
deprecated==1.2.14
    # via pygithub
distlib==0.3.8
    # via virtualenv
et-xmlfile==1.1.0
    # via openpyxl
filelock==3.13.4
    # via virtualenv
frictionless==5.16.1
    # via hdx-python-utilities
hdx-python-country==3.7.0
    # via hapi-pipelines (pyproject.toml)
hdx-python-utilities==3.6.7
    # via hdx-python-country
humanize==4.9.0
    # via frictionless
identify==2.5.35
    # via pre-commit
idna==3.7
    # via requests
ijson==3.2.3
    # via hdx-python-utilities
iniconfig==2.0.0
    # via pytest
isodate==0.6.1
    # via frictionless
jinja2==3.1.3
    # via frictionless
jsonlines==4.0.0
    # via hdx-python-utilities
jsonpath-ng==1.6.1
    # via libhxl
jsonschema==4.17.3
    # via
    #   frictionless
    #   tableschema-to-template
libhxl==5.2.1
    # via
    #   hapi-pipelines (pyproject.toml)
    #   hdx-python-country
loguru==0.7.2
    # via hdx-python-utilities
markdown-it-py==3.0.0
    # via rich
marko==2.0.3
    # via frictionless
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
nodeenv==1.8.0
    # via pre-commit
openpyxl==3.1.2
    # via hdx-python-utilities
packaging==24.0
    # via pytest
petl==1.7.15
    # via frictionless
platformdirs==4.2.0
    # via virtualenv
pluggy==1.4.0
    # via pytest
ply==3.11
    # via
    #   jsonpath-ng
    #   libhxl
pre-commit==3.7.0
    # via hapi-pipelines (pyproject.toml)
pycparser==2.22
    # via cffi
pydantic==2.7.0
    # via frictionless
pydantic-core==2.18.1
    # via pydantic
pygithub==2.3.0
    # via hapi-pipelines (pyproject.toml)
pygments==2.17.2
    # via rich
pyjwt[crypto]==2.8.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyphonetics==0.5.3
    # via hdx-python-country
pyrsistent==0.20.0
    # via jsonschema
pytest==8.1.1
    # via
    #   hapi-pipelines (pyproject.toml)
    #   pytest-cov
pytest-cov==5.0.0
    # via hapi-pipelines (pyproject.toml)
python-dateutil==2.8.2
    # via
    #   frictionless
    #   hdx-python-utilities
    #   libhxl
python-io-wrapper==0.3.1
    # via libhxl
python-slugify==8.0.4
    # via frictionless
pyyaml==6.0.1
    # via
    #   frictionless
    #   pre-commit
    #   tableschema-to-template
ratelimit==2.2.1
    # via hdx-python-utilities
requests==2.31.0
    # via
    #   frictionless
    #   libhxl
    #   pygithub
    #   requests-file
requests-file==2.0.0
    # via hdx-python-utilities
rfc3986==2.0.0
    # via frictionless
rich==13.7.1
    # via typer
ruamel-yaml==0.18.6
    # via hdx-python-utilities
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
shellingham==1.5.4
    # via typer
simpleeval==0.9.13
    # via frictionless
six==1.16.0
    # via
    #   isodate
    #   python-dateutil
stringcase==1.2.0
    # via frictionless
structlog==24.1.0
    # via libhxl
tableschema-to-template==0.0.13
    # via hdx-python-utilities
tabulate==0.9.0
    # via frictionless
text-unidecode==1.3
    # via python-slugify
typer[all]==0.12.3
    # via frictionless
typing-extensions==4.11.0
    # via
    #   frictionless
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   typer
unidecode==1.3.8
    # via
    #   libhxl
    #   pyphonetics
urllib3==2.2.1
    # via
    #   libhxl
    #   pygithub
    #   requests
validators==0.28.0
    # via frictionless
virtualenv==20.25.1
    # via pre-commit
wheel==0.43.0
    # via libhxl
wrapt==1.16.0
    # via deprecated
xlrd==2.0.1
    # via hdx-python-utilities
xlrd3==1.1.0
    # via libhxl
xlsxwriter==3.2.0
    # via tableschema-to-template
xlwt==1.3.0
    # via hdx-python-utilities

# The following packages are considered to be unsafe in a requirements file:
# setuptools
